Tidy up: Moved NodeExtensions to Extensions namespace.
