+++ Mjml.Net/Components/Body/DividerComponent.cs
@@ -1,4 +1,5 @@
 ﻿#pragma warning disable SA1119 // Statement should not use unnecessary parenthesis
+using Mjml.Net.Extensions;
 
 namespace Mjml.Net.Components.Body
 {
