+++ Mjml.Net/Components/Body/HeroComponent.cs
@@ -1,4 +1,5 @@
 ﻿#pragma warning disable RECS0018 // Comparison of floating point numbers with equality operator
+using Mjml.Net.Extensions;
 
 namespace Mjml.Net.Components.Body
 {
