+++ Mjml.Net/Components/ContainerWidth.cs
@@ -1,4 +1,6 @@
-﻿namespace Mjml.Net.Components
+﻿using Mjml.Net.Extensions;
+
+namespace Mjml.Net.Components
 {
 #pragma warning disable SA1313 // Parameter names should begin with lower-case letter
     public sealed record ContainerWidth(double Value, string String, string StringWithUnit)
