+++ Mjml.Net/Extensions/NodeExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System.Globalization;
+using Mjml.Net.Components;
 
-namespace Mjml.Net.Components
+namespace Mjml.Net.Extensions
 {
     public static class NodeExtensions
     {
